move search to inventory from middleware
